Footer content updates, finished About page
Completed content for About page. For a minute, some of the text was
appearing bold randomly with no markup, but it went away.

Made some very minor updates to footer text. Added a period to icon
attribution sentence and removed the day because I didn't feel like
editing every time. I need to come up with a script to automatically
update all of the common elements simultaneously.